notification-service/requirements.txt
amqp==5.1.1
asgiref==3.5.2
astor==0.8.1
async-timeout==4.0.2
attrs==21.4.0
bandit==1.7.4
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
darglint==1.8.1
dataclasses-json==0.5.7
Deprecated==1.2.13
Django==4.0.6
django-celery-beat==2.3.0
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
django-timezone-field==5.0
docutils==0.19
eradicate==2.1.0
Faker==13.15.0
flake8==4.0.1
flake8-bandit==3.0.0
flake8-broken-line==0.4.0
flake8-bugbear==22.7.1
flake8-commas==2.1.0
flake8-comprehensions==3.10.0
flake8-debugger==4.1.2
flake8-docstrings==1.6.0
flake8-eradicate==1.2.1
flake8-html==0.4.2
flake8-isort==4.1.1
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pbr==5.9.0
pep8-naming==0.12.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.1
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.2
pytest-celery==0.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
restructuredtext-lint==1.4.0
sendgrid==6.9.7
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
split-settings==1.0.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stevedore==4.0.0
testfixtures==6.18.5
tomli==2.0.1
types-cryptography==3.3.21
types-pyOpenSSL==22.0.10
types-pytz==2022.1.1
types-PyYAML==6.0.10
types-requests==2.28.5
types-setuptools==63.2.2
types-urllib3==1.26.17
typing-inspect==0.7.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wemake-python-styleguide==0.16.1
wrapt==1.14.1
